initial checkin the separating test cases for different browse views
